refactor: change author_id and genre_id to BigSerial for Long support

+++ sql/01-init.sql
@@ -1,12 +1,12 @@
 -- Authors table
 CREATE TABLE IF NOT EXISTS Authors (
-    author_id SERIAL PRIMARY KEY,
+    author_id BIGSERIAL PRIMARY KEY,
     author_name TEXT
 );
 
 -- Genres table
 CREATE TABLE IF NOT EXISTS Genres (
-    genre_id SERIAL PRIMARY KEY,
+    genre_id BIGSERIAL PRIMARY KEY,
     genre_name TEXT
 );
 
@@ -19,13 +19,13 @@ CREATE TABLE IF NOT EXISTS Books (
 -- Book_Author table (many-to-many relationship)
 CREATE TABLE IF NOT EXISTS Book_Author (
     isbn TEXT REFERENCES Books(isbn),
-    author_id INTEGER REFERENCES Authors(author_id),
+    author_id BIGINT REFERENCES Authors(author_id),
     PRIMARY KEY (isbn, author_id)
 );
 
 -- Book_Genre table (many-to-many relationship)
 CREATE TABLE IF NOT EXISTS Book_Genre (
     isbn TEXT REFERENCES Books(isbn),
-    genre_id INTEGER REFERENCES Genres(genre_id),
+    genre_id BIGINT REFERENCES Genres(genre_id),
     PRIMARY KEY (isbn, genre_id)
 );
